Use Active Storage to render attachments
This way we fix a bug we mentioned in commit XXXXX that caused links to
documents to be broken when editing their title because the title was
used to generate the URL of the document.

Note we're still using Paperclip to render cached attachments because
this is the only case where we store files with just Paperclip and not
Active Storage.