Add short titles to SDG targets TODO: check crowdin and add Spanish translations
Since targets didn't have a title but only a long description, every
form allowing to select targets was pretty much unusable: we either
displayed just the code or the whole description.

Now, with a concise title, it's easier to find and select the desired
target.

The titles have been copied from The Global Goals page [1].

Note we're using the `summary` I18n key for the `title` method and the
`title` I18n key for the `official_title` method. From the clean code
point of view, renaming the `title` I18n key to something else (like
`description`) would be better. However, renaming I18n keys is quite a
pain due to the way we manage translations with Crowdin.

[1] https://www.globalgoals.org/