Add information about budget actions
Both of these actions benefit from some kind of hint.

The "calculate winners" action informs administrators that results won't
be immediately available unless the "show results" option is enabled.
We're also adding emphasis to some text when results will be available
but not when they won't. This way users will notice the difference
between the scenarios, since users familiar with the interface won't
read the text.

The delete action was redirecting with an error message when the budget
couldn't be deleted; IMHO it's better to disable it and inform
administrators why it's disabled. Alternatively we could remove the
button completely; however, users might be looking for a way to delete a
budget and wouldn't find any hint about it.

We're now removing the "Delete" action from the budgets index table,
since most of the time it isn't possible to delete a budget and so the
action takes up space and we get little gain in return. We could keep
the "Delete" icon just for budgets which can be deleted; however, the
alignment of the table rows would suffer, making it harder to find the
intended action.